perf: implement source freeing algorithm based on note end time
The `AudioSourcePool` is backed by an AVL tree that tracks the end time of playing notes. When a source needs to be freed, it finds the note that's closest to ending, which corresponds to the min key of the tree.

This inrastructure makes it possible to implement other heuristics for freeing sounds - for instance, something based on velocity (lower velocity sounds get cut first), or based on note start time (not end time).

However, the current implementation doesn't sound very good, likely since it penalizes short/high notes too much, so they almost don't come through. So it will be undone on the commit following this one in favor of a simple queue, removing the oldest notes to start playing, which in general offers better sounding results.